requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.3.0
Babel==2.15.0
bangla==0.0.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coqui-tts==0.24.1
coqui-tts-trainer==0.1.4
ctranslate2==4.3.1
cutlet==0.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
face-alignment==1.4.1
fastapi==0.110.3
faster-whisper==1.0.2
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fugashi==1.3.2
g2pkk==0.1.2
gradio==4.13.0
gradio_client==0.8.0
grpcio==1.65.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
imageio==2.37.0
importlib_resources==6.4.0
inflect==7.3.1
jaconv==0.4.0
jamo==0.4.1
jax==0.5.3
jaxlib==0.5.3
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
lws==1.2.8
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mecab-python3==1.0.9
mediapipe==0.10.21
ml_dtypes==0.5.1
mojimoji==0.0.13
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==3.4.2
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-contrib-python==4.6.0.66
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.4
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pyreadline3==3.4.1
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
resampy==0.4.3
rich==13.7.1
rpds-py==0.19.1
safetensors==0.4.3
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2025.3.13
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.4
transformers==4.40.2
typeguard==4.3.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
unidic-lite==1.0.8
urllib3==2.2.2
uvicorn==0.30.3
wasabi==1.1.3
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.3
yarl==1.9.4
